rename files to match class names
