+++ src/DetailsView/components/no-content-available.tsx
