+++ src/DetailsView/details-view-container.tsx
@@ -6,6 +6,7 @@ import { CardSelectionStoreData } from 'common/types/store-data/card-selection-s
 import { ISelection } from 'office-ui-fabric-react/lib/DetailsList';
 import { Spinner, SpinnerSize } from 'office-ui-fabric-react/lib/Spinner';
 import * as React from 'react';
+
 import { ThemeDeps } from '../common/components/theme';
 import { withStoreSubscription, WithStoreSubscriptionDeps } from '../common/components/with-store-subscription';
 import { VisualizationConfigurationFactory } from '../common/configs/visualization-configuration-factory';
@@ -30,8 +31,8 @@ import { DetailsRightPanelConfiguration, GetDetailsRightPanelConfiguration } fro
 import { GetDetailsSwitcherNavConfiguration } from './components/details-view-switcher-nav';
 import { Header, HeaderDeps } from './components/header';
 import { IssuesTableHandler } from './components/issues-table-handler';
+import { NoContentAvailable } from './components/no-content-available';
 import { TargetChangeDialogDeps } from './components/target-change-dialog';
-import { NoContentAvailable } from './components/target-page-closed-view';
 import { DetailsViewBody, DetailsViewBodyDeps } from './details-view-body';
 import { AssessmentInstanceTableHandler } from './handlers/assessment-instance-table-handler';
 import { DetailsViewToggleClickHandlerFactory } from './handlers/details-view-toggle-click-handler-factory';
