+++ src/tests/unit/tests/DetailsView/components/no-content-available.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { NoContentAvailable } from 'DetailsView/components/target-page-closed-view';
+import { NoContentAvailable } from 'DetailsView/components/no-content-available';
 import { shallow } from 'enzyme';
 import * as React from 'react';
